201410232237
Originally checked in by Hujolopter\Sebastian on 2014-10-23 22:37:53.08 as Changeset 55

git-tfs-id: [https://infovision.visualstudio.com/DefaultCollection]$/RoughSets;C91

+++ RoughSets/Infovision.Datamining.RoughSet/PermutationGeneratorFieldObject.cs
@@ -77,7 +77,7 @@ namespace Infovision.Datamining.Roughset
 
                 if (fieldList.Count > 0 && objectList.Count > 0)
                 {
-                    if (this.fieldSelectionRatio >= (double)1
+                    if (this.fieldSelectionRatio >= 1.0
                         || RandomSingleton.Random.NextDouble() < this.fieldSelectionRatio)
                     {
                         element = this.GetAndRemoveListElement<int>(fieldList);
